requirements.txt
jsonref 
PyYAML